+++ src/main/java/kitchenpos/application/tablegroup/TableGroupService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.tablegroup;
 
 import java.util.Arrays;
 import java.util.List;
